+++ kernel/sys/fal/fal-internals.h
@@ -33,6 +33,6 @@ ssize_t sys_write(stack_frame_t* regs, int fd, void* buf, size_t count);
 int sys_close(stack_frame_t* regs, int fd);
 uint64_t sys_getcwd(stack_frame_t* regs, const char* buf, size_t size);
 int sys_chdir(stack_frame_t* regs, const char* path);
-ssize_t sys_getdents(stack_frame_t* regs, int fd, void* dirp, size_t count);
+long sys_getdents(stack_frame_t* regs, int fd, void* dirp, unsigned int count);
 
 #endif //FAL_INTERNALS_H
